+++ script.js
@@ -1,45 +1,28 @@
-( () =>{ 
+import checkComplete from "./components/checkComplete.js";
+import deleteIcon from "./components/deleteIcon.js";
 
-    const btn = document.querySelector('[data-form-btn]');
+const btn = document.querySelector('[data-form-btn]');
 
-const createTask = (evento) => {
-    evento.preventDefault();
-    const input = document.querySelector('[data-form-input]');
-    const value = input.value;
-    const list = document.querySelector('[data-list]');
-    const task = document.createElement('li');
-    task.classList.add('card');
-    input.value = '';
-    //backticks
-    const taskContent = document.createElement('div');
-    
-    const titleTask = document.createElement('span');
-    titleTask.classList.add('task');
-    titleTask.innerText = value;
-    taskContent.appendChild(checkComplete());
-    taskContent.appendChild(titleTask);
-    const content = `
-        <span class="task">${value}</span>
-    <i class="fas fa-trash-alt trashIcon icon"></i>`
-    // task.innerHTML = content;
-    task.appendChild(taskContent);
-    list.appendChild(task);
-};
+    const createTask = (evento) => {    
+        evento.preventDefault();
+        const input = document.querySelector('[data-form-input]');
+        const value = input.value;
+        const list = document.querySelector('[data-list]');
+        const task = document.createElement('li');
+        task.classList.add('card');
+        input.value = '';
+        //backticks
+        const taskContent = document.createElement('div');
+        const titleTask = document.createElement('span');
+        titleTask.classList.add('task');
+        titleTask.innerText = value;
+        taskContent.appendChild(checkComplete());
+        taskContent.appendChild(titleTask);
+        // task.innerHTML = content;
+        task.appendChild(taskContent);
+        task.appendChild(deleteIcon());
+        list.appendChild(task);
+    };
 
-//Arrow function o funciones anonimas
-btn.addEventListener('click', createTask);
-
-const checkComplete = () => {
-    const i = document.createElement('i');
-    i.classList.add('far', 'fa-check-square', 'icon' );
-    i.addEventListener("click", completeTask);
-    return i;
-};
-//Immediately invoked fuction expression (IIFE)
-const completeTask = (event) => {
-    const element = event.target;
-    element.classList.toggle('fas');
-    element.classList.toggle('completeIcon')
-    element.classList.toggle('far');
-};
-})();
+    //Arrow function o funciones anonimas
+    btn.addEventListener('click', createTask);
